requirements.txt
streamlit>=1.7.0
tensorflow-hub==0.12.0
tensorflow-cpu ==2.5.0